--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,7 +12,7 @@ repository = "https://github.com/pop-os/cosmic-text"
 fontdb = { version = "0.13.0", default-features = false }
 libm = "0.2.6"
 log = "0.4.17"
-ouroboros = { version = "0.15.5", default-features = false }
+aliasable = "0.1.3"
 rustybuzz = { version = "0.7.0", default-features = false, features = ["libm"] }
 swash = { version = "0.1.6", optional = true }
 syntect = { version = "5.0.0", optional = true }
@@ -35,7 +35,6 @@ no_std = [
 std = [
   "fontdb/memmap",
   "fontdb/std",
-  "ouroboros/std",
   "rustybuzz/std",
   "sys-locale",
   "unicode-bidi/std",