@bjorn3 2023-06-08
https://github.com/axelf4/unicode-linebreak/pull/8 also helps with compilation time.