@bjorn3 2023-06-08
Looks like there may be an easy win for swash: https://github.com/dfrg/swash/issues/37#issuecomment-1583035890